// apps/indexer/src/database/schemas/public/DeadRareListing.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

import { type ColumnType, type Selectable, type Insertable, type Updateable } from 'kysely';

/** Identifier type for public.DeadRareListing */
export type DeadRareListingId = string & { __brand: 'DeadRareListingId' };

/** Represents the table public.DeadRareListing */
export default interface DeadRareListingTable {
  id: ColumnType<DeadRareListingId, DeadRareListingId | undefined, DeadRareListingId>;

  price: ColumnType<bigint, bigint, bigint>;

  listingId: ColumnType<bigint, bigint, bigint>;

  collectionAddress: ColumnType<string, string, string>;

  tokenAddress: ColumnType<string, string, string>;

  seller: ColumnType<string, string, string>;

  buyer: ColumnType<string | null, string | null, string | null>;

  listingAddress: ColumnType<string, string, string>;

  listedAt: ColumnType<Date, Date | string, Date | string>;

  listedTransaction: ColumnType<string, string, string>;

  soldAt: ColumnType<Date | null, Date | string | null, Date | string | null>;

  soldTransaction: ColumnType<string | null, string | null, string | null>;

  unlistedAt: ColumnType<Date | null, Date | string | null, Date | string | null>;

  unlistedTransaction: ColumnType<string | null, string | null, string | null>;
}

export type DeadRareListing = Selectable<DeadRareListingTable>;

export type NewDeadRareListing = Insertable<DeadRareListingTable>;

export type DeadRareListingUpdate = Updateable<DeadRareListingTable>;